move attls handler to the common module
