+++ apiml-tomcat-common/build.gradle
@@ -1,6 +1,10 @@
 dependencies {
+
+    api project(':common-service-core')
     api libs.spring.boot.starter.web
     api libs.zowe.zos.utils
+    implementation libs.netty.reactor.http
+    implementation libs.apache.commons.lang3
 
     testImplementation libs.spring.boot.starter.test
